requirements.txt
# ── Data acquisition ──────────────────────────────────────────────────────
yfinance>=0.2.36
requests>=2.31.0

# ── Data manipulation ─────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
openpyxl>=3.1.0

# ── Technical analysis ────────────────────────────────────────────────────
pandas-ta>=0.3.14b

# ── Machine learning ──────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0
hmmlearn>=0.3.0
shap>=0.44.0
optuna>=3.4.0
joblib>=1.3.0

# ── Streamlit app ─────────────────────────────────────────────────────────
streamlit>=1.30.0
plotly>=5.18.0

# ── Scheduler ─────────────────────────────────────────────────────────────
APScheduler>=3.10.0
pytz>=2024.1

# ── Configuration & environment ───────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ── Utilities ─────────────────────────────────────────────────────────────
tqdm>=4.66.0

# ── Testing ───────────────────────────────────────────────────────────────
pytest>=7.4.0
